+++ test/integration/testHappyPathTwilio.js
@@ -75,9 +75,6 @@ describe('happy path Twilio integration test: responder responds right away and 
     // Expect to log the Twilio ID
     expect(helpers.log.getCall(0)).to.be.calledWithMatch('Sent by Twilio:')
 
-    // Expect the state to change to STARTED
-    expect(this.braveAlerter.alertSessionChangedCallback).to.be.calledWith(new AlertSession(sessionId, CHATBOT_STATE.STARTED))
-
     this.currentAlertSession.alertState = CHATBOT_STATE.STARTED
 
     // Responder replies 'Ok'